Spectral models for Gammapy
No, really ... this time it's working!
