+++ gammapy/spectrum/__init__.py
@@ -10,12 +10,10 @@ from .fitter import *
 from .fitting_utils import *
 from .flux_point import *
 from .isrf import *
-from .models import *
 from .powerlaw import *
 from .sed import *
 from .sherpa_chi2asym import *
 from .utils import *
-#from .counts_spectrum import *
 from .spectrum_extraction import *
 from .spectrum_grouping import *
 from .spectrum_fit import *
