+++ gammapy/spectrum/tests/test_models.py
@@ -1,0 +1,4 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+from __future__ import absolute_import, division, print_function, unicode_literals
+from ..models import PowerLaw, ExponentialCutoffPowerLaw
+
